pyproject.toml
[project]
name = "geoconfig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "taraaubrey", email = "tara.a.forstner@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "strictyaml>=1.7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
